+++ src/libstddjb/selfpipe_read.c
@@ -16,7 +16,7 @@ int selfpipe_read (void)
 {
   struct signalfd_siginfo buf ;
   ssize_t r = sanitize_read(fd_read(selfpipe_fd, (char *)&buf, sizeof(struct signalfd_siginfo))) ;
-  return (r <= 0) ? r : (int)buf.ssi_signo ;
+  return (r <= 0) ? r : buf.ssi_signo ;
 }
       
 #else
@@ -25,7 +25,7 @@ int selfpipe_read (void)
 {
   char c ;
   ssize_t r = sanitize_read((fd_read(selfpipe_fd, &c, 1))) ;
-  return (r <= 0) ? r : (int)c ;
+  return (r <= 0) ? r : c ;
 }
 
 #endif
